+++ ajwah_bloc/test/counterState.dart
@@ -27,7 +27,7 @@ class CounterState extends BaseState<CounterModel> {
   CounterState() : super(name: 'counter', initialState: CounterModel.init());
 
   Stream<CounterModel> mapActionToState(
-      CounterModel state, Action action) async* {
+      CounterModel state, Action action, Store store) async* {
     switch (action.type) {
       case ActionTypes.Inc:
         yield CounterModel.countData(state.count + 1);
@@ -40,7 +40,7 @@ class CounterState extends BaseState<CounterModel> {
         break;
 
       default:
-        yield state;
+        yield getState(store);
     }
   }
 }