+++ ajwah_bloc_examples/lib/main.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:ajwah_bloc/ajwah_bloc.dart' as store;
 
 void main() {
-  createStore(states: [CounterState()]);
+  createStore(states: [CounterState()], enableGlobalApi: true);
 
   runApp(MyApp());
 }
@@ -162,7 +162,7 @@ class CounterState extends BaseState<CounterModel> {
   CounterState() : super(name: 'counter', initialState: CounterModel.init());
 
   Stream<CounterModel> mapActionToState(
-      CounterModel state, store.Action action) async* {
+      CounterModel state, store.Action action, Store store) async* {
     switch (action.type) {
       case 'Inc':
         yield state.copyWith(count: state.count + 1, isLoading: false);
@@ -173,10 +173,10 @@ class CounterState extends BaseState<CounterModel> {
       case 'AsyncInc':
         yield state.copyWith(isLoading: true);
         await Future.delayed(Duration(seconds: 1));
-        dispatch('Inc');
+        store.dispatch('Inc');
         break;
       default:
-        yield latestState(this);
+        yield getState(store);
     }
   }
 }