+++ ajwah_bloc_examples/test/widget_test.dart
@@ -17,6 +17,9 @@ void main() {
   setUpAll(() {
     store = createStore(states: [CounterState()]);
   });
+  tearDownAll(() {
+    store.dispose();
+  });
   ajwahTest(
     "initial counter state data[10,flase]",
     build: () => store.select('counter').map((event) => event.toString()),
@@ -25,14 +28,14 @@ void main() {
   ajwahTest(
     "check increment[11,false]",
     build: () => store.select('counter').map((event) => event.toString()),
-    act: () => store.dispatch(Action(type: 'Inc')),
+    act: () => store.dispatcH(Action(type: 'Inc')),
     skip: 1,
     expect: ['{coun:11, isLoading:false}'],
   );
   ajwahTest(
     "check decrement[10,false]",
     build: () => store.select('counter').map((event) => event.toString()),
-    act: () => store.dispatch(Action(type: 'Dec')),
+    act: () => store.dispatcH(Action(type: 'Dec')),
     skip: 1,
     expect: ['{coun:10, isLoading:false}'],
   );
@@ -40,7 +43,7 @@ void main() {
   ajwahTest(
     "check async increment[11,false]",
     build: () => store.select('counter').map((event) => event.toString()),
-    act: () => store.dispatch(Action(type: 'AsyncInc')),
+    act: () => store.dispatcH(Action(type: 'AsyncInc')),
     wait: const Duration(milliseconds: 1000),
     skip: 1,
     expect: [
@@ -52,7 +55,7 @@ void main() {
   ajwahTest(
     "check decrement[10,false]",
     build: () => store.select('counter').map((event) => event.toString()),
-    act: () => store.dispatch(Action(type: 'Dec')),
+    act: () => store.dispatcH(Action(type: 'Dec')),
     skip: 1,
     expect: ['{coun:10, isLoading:false}'],
   );